refactor(LSG): move classes and types around
