+++ packages/live-status-gateway/src/collections/rundownHandler.ts
@@ -1,6 +1,7 @@
 import { Logger } from 'winston'
 import { CoreHandler } from '../coreHandler'
-import { Collection, PickArr, PublicationCollection } from '../wsHandler'
+import { Collection } from '../wsHandler'
+import { PublicationCollection } from '../publicationCollection'
 import { DBRundownPlaylist } from '@sofie-automation/corelib/dist/dataModel/RundownPlaylist'
 import { DBRundown } from '@sofie-automation/corelib/dist/dataModel/Rundown'
 import { RundownId, RundownPlaylistId } from '@sofie-automation/corelib/dist/dataModel/Ids'
@@ -9,9 +10,10 @@ import { RundownsHandler } from './rundownsHandler'
 import { CorelibPubSub } from '@sofie-automation/corelib/dist/pubsub'
 import { unprotectString } from '@sofie-automation/server-core-integration'
 import { CollectionHandlers } from '../liveStatusServer'
+import { PickKeys } from '@sofie-automation/shared-lib/dist/lib/types'
 
 const PLAYLIST_KEYS = ['_id', 'currentPartInfo', 'nextPartInfo'] as const
-type Playlist = PickArr<DBRundownPlaylist, typeof PLAYLIST_KEYS>
+type Playlist = PickKeys<DBRundownPlaylist, typeof PLAYLIST_KEYS>
 
 export class RundownHandler
 	extends PublicationCollection<DBRundown, CorelibPubSub.rundownsInPlaylists, CollectionName.Rundowns>
