+++ packages/live-status-gateway/src/collections/segmentsHandler.ts
@@ -1,6 +1,7 @@
 import { Logger } from 'winston'
 import { CoreHandler } from '../coreHandler'
-import { CollectionBase, Collection } from '../wsHandler'
+import { Collection } from '../wsHandler'
+import { CollectionBase } from '../collectionBase'
 import { DBSegment } from '@sofie-automation/corelib/dist/dataModel/Segment'
 import * as _ from 'underscore'
 import { CollectionName } from '@sofie-automation/corelib/dist/dataModel/Collections'
